third_party/WebKit/LayoutTests/http/tests/inspector/security/security-blocked-mixed-content-and-malicious.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../security-test.js"></script>
<script>
function test()
{
    /** @type {!Protocol.Security.InsecureContentStatus} */
    var insecureContentStatus = { ranMixedContent: false, displayedMixedContent: false, ranContentWithCertErrors: false, displayedContentWithCertErrors: false, ranInsecureContentStyle: Protocol.Security.SecurityState.Insecure, displayedInsecureContentStyle: Protocol.Security.SecurityState.Neutral };
    InspectorTest.mainTarget.model(Security.SecurityModel).dispatchEventToListeners(Security.SecurityModel.Events.SecurityStateChanged, new Security.PageSecurityState(Protocol.Security.SecurityState.Secure, true, [], insecureContentStatus, "Test: Summary Override Text"));

    var request = new SDK.NetworkRequest(InspectorTest.mainTarget, 0, "http://foo.test", "https://foo.test", 0, 0, null);
    request.setBlockedReason(Protocol.Network.BlockedReason.MixedContent);
    request.mixedContentType = "blockable";
    InspectorTest.dispatchRequestFinished(request);

    InspectorTest.dumpDeepInnerHTML(Security.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-summary-text")[0]);

    var explanations = Security.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-explanation");
    for (var i = 0; i < explanations.length; i++)
        InspectorTest.dumpDeepInnerHTML(explanations[i]);

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests active mixed content blocking in the security panel when page is classified as malicious.</p>
</body>
</html>
