third_party/WebKit/LayoutTests/http/tests/inspector/security/security-blocked-mixed-content-and-malicious-expected.txt

Tests active mixed content blocking in the security panel when page is classified as malicious.

<DIV class=security-summary-text >
Test: Summary Override Text
</DIV>
<DIV class=security-explanation security-explanation-secure >
    <DIV class=security-property security-property-secure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Secure Resources
        </DIV>
        <DIV >
All resources on this page are served securely.
        </DIV>
    </DIV>
</DIV>
<DIV class=security-explanation security-explanation-info >
    <DIV class=security-property security-property-info >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Blocked mixed content
        </DIV>
        <DIV >
Your page requested non-secure resources that were blocked.
        </DIV>
        <DIV class=security-mixed-content link >
View 1 request in Network Panel
        </DIV>
    </DIV>
</DIV>
